Enabling LLama model on CPU

+++ auto_gptq/nn_modules/qlinear/qlinear_cuda.py
@@ -267,10 +267,10 @@ class QuantLinear(nn.Module):
                     g_idx_i = self.g_idx[i*num_dim:(i+1)*num_dim]
                     weights.append(scale_i[g_idx_i.long()] * (weight_i - zeros_i[g_idx_i.long()]))
                 weights = torch.cat(weights,dim=1)
-            out = torch.matmul(x, weights)
+            out = torch.matmul(x.float(), weights.float())
         out = out.reshape(out_shape)
         out = out + self.bias if self.bias is not None else out
-        return out
+        return out.half()
 
 
 __all__ = ["QuantLinear"]
